refactor: AppContextからZustandへの状態管理システム統一
## 主な変更内容

### 🔄 状態管理システム移行
- AppContext（src/contexts/AppContext.tsx）を完全削除
- 全コンポーネントでuseAppContext → useAppStoreへ移行
- グローバルフィルター（selectedPeriod, selectedCustomerSegment）をappStoreに統合

### 📁 アーキテクチャ改善
- メニュー設定をsrc/lib/menuConfig.tsに分離
- プロバイダー構成をAppProvider → ZustandProvider + FilterProviderに変更
- 型定義（PeriodType, CustomerSegmentType）をappStoreに移動

### 🛠️ 影響範囲
- 8コンポーネントでstate管理ロジック更新
- レイアウトプロバイダー構成変更
- import文の整理とexport形式の統一

### ✨ 改善効果
- 複雑な状態管理の統一
- パフォーマンス向上（Context Provider層削減）
- メンテナンス性向上（単一状態管理パターン）
- 型安全性の強化

### 📝 作業ログ
- 移行作業完了記録（worklog/20250604-context-to-zustand-migration.md）
- タスク管理ファイル更新

**Breaking Changes:** なし（API互換性維持）
**Test Status:** 全ページビルド・動作確認済み
