+++ worklog/tasks/main-todo.md
@@ -2,10 +2,10 @@
 
 ## 🎯 現在のステータス
 
-- **プロジェクト状況**: 開発段階（**技術的負債改善プロジェクト開始**） ⭐NEW
-- **最終更新**: 2025年5月25日
-- **主要課題**: 技術的負債解消、API切り替え準備、共通コンポーネント化
-- **現在フェーズ**: **技術的負債分析完了・改善実装フェーズ** ⭐NEW
+- **プロジェクト状況**: 開発段階（**状態管理システム統一完了**） ⭐NEW
+- **最終更新**: 2025年6月4日
+- **主要課題**: ✅ 技術的負債解消90%完了、状態管理統一完了、品質向上フェーズ移行
+- **現在フェーズ**: **技術的負債解消→品質最適化フェーズ** ⭐NEW
 - **学習プロジェクト**: React/Next.js/TypeScript実践学習（**アーキテクチャ改善マスター**）
 - **学習戦略**: 「分析→基盤整備→最適化→品質向上」パターン確立
 - **データ戦略**: **APIデータ/モックデータ並行使用体制確立**
@@ -236,6 +236,6 @@
 
 ---
 
-*最終更新: 2025年5月25日（技術的負債分析統合）*
+*最終更新: 2025年6月4日（状態管理システム統一完了）*
 *作成者: AI Assistant*
 *次回レビュー: 技術的負債Phase 1完了時*
